Split ESP flash contract types
Extract ESP flash state, payload shapes, dataclasses, and protocols into focused neighbor modules, move serial discovery and subprocess execution helpers out of the orchestrator, and keep esp_flash_manager.py as the flash state machine.
