+++ apps/server/tests/use_cases/updates/test_esp_flash_phase_coherence.py
@@ -16,11 +16,8 @@ from __future__ import annotations
 
 from pathlib import Path
 
-from vibesensor.use_cases.updates.esp_flash_manager import (
-    EspFlashManager,
-    EspFlashState,
-    EspFlashStatus,
-)
+from vibesensor.use_cases.updates.esp_flash_manager import EspFlashManager
+from vibesensor.use_cases.updates.esp_flash_types import EspFlashState, EspFlashStatus
 
 # ── Helpers ───────────────────────────────────────────────────────────────────
