+++ docs/ai/repo-map.md
@@ -40,7 +40,7 @@ Scope: single source of truth for detailed file layout, entry points, package st
 - `use_cases/diagnostics/`: post-stop diagnostics pipeline. See `docs/analysis_pipeline.md` for the full data flow, pipeline steps, and module map. `run_data_preparation.py` owns `PreparedRunData` plus run timing/speed/phase/sensor preparation, `findings.py` owns reference-check and peak/order orchestration, `peak_binning.py` owns persistent-peak accumulation/scoring, `peak_table.py` owns internal peak-table row ranking/statistics, `signal_aggregation.py` owns speed/location summary aggregation, `math_utils.py` owns generic statistics/correlation helpers, `speed_profile_helpers.py` owns speed-profile assembly helpers, `order_analysis.py` now owns the core order-matching/scoring/assembly primitives, `order_pipeline.py` owns `OrderAnalysisSession` plus the `_build_order_findings()`/multi-location orchestration layer above those primitives, and `order_heuristics.py` owns the pure diffuse-excitation, localization-override, and engine-alias heuristics; `top_cause_selection.py` delegates classification and ranking to the domain `Finding`; `location_analysis.py` owns the location-analysis pipeline and `LocationAnalysisResult` typed return; `plots.py` now keeps plot-series/FFT/spectrogram orchestration over diagnostics-local value objects, `summary_builder.py` focuses on `RunAnalysis` orchestration and domain/app summary assembly inputs, and `_types.py` owns diagnostics-local helper/value types such as `AccelStatistics`, speed-breakdown rows, and plot/spectrogram value objects. Boundary summary payloads live in `shared/boundaries/analysis_payload.py`, while `shared/boundaries/summary_serialization/` is the dedicated domain/app → payload package: `_contracts.py` owns structural typing protocols, `_findings.py` owns finding annotation/serialization, `_plots.py` owns plot and table serialization, and `_summary.py` owns top-level summary assembly. History-use-case callers still receive persisted-summary projection through injected adapter wiring instead of importing that projector here. The package `__init__.py` re-exports only high-level analysis entrypoints; PDF rendering types/helpers now live under `adapters/pdf/`.
 - `use_cases/history/`: run query/delete, PDF report generation, CSV/ZIP exports, and history-facing helper orchestration above persistence ports. `helpers.py` owns the local `HistoryRecord` TypedDict plus generic history workflow helpers, while `report_interpretation.py` owns pure report-domain interpretation consumed by the PDF mapper.
 - `use_cases/run/`: recording pipeline orchestration; `lifecycle_state.py` owns `RunLifecycleState` plus the active-run snapshot, `persistence_writer.py` owns history-write coordination/retry bookkeeping, `sample_flush.py` owns sample-building/flush decisions and auto-stop checks, `status_reporting.py` owns the status/health payload helpers, `logger.py` owns the thin `RunRecorder` orchestrator, `post_analysis.py` owns the background analysis queue above the injected run-persistence/analysis/error boundary, and `sample_builder.py` owns pure sample-building helpers.
-- `use_cases/updates/`: wheel-based updater workflow orchestration, runtime prerequisite validation, firmware cache, ESP flashing, release discovery, install, rollback, runner, Wi-Fi, and status tracking. `firmware_cache.py` now stays the thin public cache/CLI facade, `firmware_types.py` owns firmware cache/release contracts, `firmware_bundle.py` owns bundle filesystem validation/extraction/metadata helpers, and `firmware_release_fetcher.py` owns GitHub firmware HTTP access.
+- `use_cases/updates/`: wheel-based updater workflow orchestration, runtime prerequisite validation, firmware cache, ESP flashing, release discovery, install, rollback, runner, Wi-Fi, and status tracking. `firmware_cache.py` stays the thin public cache/CLI facade, `firmware_types.py` owns firmware cache/release contracts, `firmware_bundle.py` owns bundle filesystem validation/extraction/metadata helpers, `firmware_release_fetcher.py` owns GitHub firmware HTTP access, `esp_flash_manager.py` owns the flash orchestration state machine, `esp_flash_types.py` owns ESP flash contracts/state/protocols, and `esp_serial.py` / `esp_flash_runner.py` own serial discovery and subprocess execution helpers.
 - `shared/`: cross-cutting typed payloads and ports (`shared/types/` and `shared/ports.py`), boundary serializers/decoders (`shared/boundaries/`), exceptions (`shared/exceptions.py`), JSON helpers (`shared/json_utils.py`), location identifiers (`shared/locations.py`), run-context helpers, and shared pure vehicle-order math (`shared/order_bands.py`). `shared/ports.py` owns the current run-history persistence, active-client lookup, latest-sample/latest-metrics, resolved-speed, and read-only settings ports used by runtime/history/recording flows; `shared/types/` stays focused on payload/model types such as `sensor_frame.py`, `health_snapshot.py`, the feature-scoped Pydantic HTTP contract package `shared/types/api_models/`, and the typed backend/API payloads; `shared/boundaries/run_log.py` owns JSONL run-log decoding/normalization shared by diagnostics and persistence; and `shared/boundaries/diagnostic_case.py` owns summary projection, typed speed/suitability decoding, and shared metadata-to-case reconstruction (`case_context_from_metadata`).
 - `domain/`: DDD-aligned domain model package. Primary domain objects live under `vibesensor/domain/`; larger supporting value objects now split into focused neighboring modules when their behavior grows beyond their parent aggregate file (for example `finding_types.py`, `finding_evidence.py`, `tire_spec.py`, and `order_reference.py`). Domain objects own classification, ranking, actionability, surfacing, lifecycle, and query logic; diagnostics use cases delegate to them. See `docs/domain-model.md` for the full domain object catalog, relationship map, and modeling rules.
 - `apps/ui/src/app/runtime/`: explicit UI runtime owners for shell/chrome state, live transport/payload application, and spectrum/chart orchestration beneath the `UiAppRuntime` composition root.
